backend/requirements.txt
fastapi
motor
pydantic
python-dotenv
uvicorn[standard]
crewai
masumi
python-multipart
httpx
requests
SQLAlchemy
psycopg2-binary

